refactor: project structure update
