+++ server/src/main/kotlin/com/appsheaven/keepup/todos/domain/TodoDatabase.kt
@@ -1,13 +1,13 @@
-package com.appsheaven.keepup.core
+package com.appsheaven.keepup.todos.domain
 
-import com.appsheaven.keepup.todos.Todo
+import com.appsheaven.keepup.todos.entities.Todo
 import com.mongodb.MongoException
 import com.mongodb.client.result.DeleteResult
 import com.mongodb.client.result.InsertOneResult
 import com.mongodb.client.result.UpdateResult
 import org.bson.types.ObjectId
 
-internal interface KeepUpDatabase {
+internal interface TodoDatabase {
     @Throws(MongoException::class)
     suspend fun insertTodo(todo: Todo): InsertOneResult
     suspend fun getTodos(page: Int, pageSize: Int): List<Todo>
