+++ server/src/test/kotlin/com/appsheaven/keepup/database/MongoDatabaseFixtures.kt
@@ -1,6 +1,6 @@
 package com.appsheaven.keepup.database
 
-import com.appsheaven.keepup.todos.Todo
+import com.appsheaven.keepup.todos.entities.Todo
 import org.bson.types.ObjectId
 
 internal object MongoDatabaseFixtures {
